Initial commit for birthday wisher email project
This python project wishes happy birthday to the person from the birthdays.csv file according to their date and month of birthday. It will fetch a random letter from letter_templates folder to prepare a letter and then it will later email that to the person. The name, email, dob of the person is taken from the bithdays.csv file.